+++ src/3rd_party/half_float/umHalf.h
@@ -243,7 +243,7 @@ class numeric_limits<HalfFloat> {
     static HalfFloat min ()
 		{return HalfFloat(0,1,0);}
     static HalfFloat max ()
-		{return HalfFloat(~0,HalfFloat::MAX_EXPONENT_VALUE-1,0);}
+		{return HalfFloat((uint16_t)~0,HalfFloat::MAX_EXPONENT_VALUE-1,0);}
     static const int radix = 2;
     static const int digits = 10;   // conservative assumption
     static const int digits10 = 2;  // conservative assumption
@@ -255,7 +255,7 @@ class numeric_limits<HalfFloat> {
     static const bool is_bounded	= true;
 
 	static const HalfFloat lowest() {
-	  return HalfFloat(~0,HalfFloat::MAX_EXPONENT_VALUE-1,~0);
+	  return HalfFloat((uint16_t)~0,HalfFloat::MAX_EXPONENT_VALUE-1,(uint16_t)~0);
 	}
 
 	// Floating point specific.
